reorg util test
