+++ test/utils/Grpc.spec.ts
@@ -1,0 +1,73 @@
+import path from 'path';
+import { InterceptingCall } from '@grpc/grpc-js';
+import { getGRPCService, getAuthInterceptor } from '../../milvus';
+// mock
+jest.mock('@grpc/grpc-js', () => {
+  const actual = jest.requireActual(`@grpc/grpc-js`);
+
+  return {
+    InterceptingCall: jest.fn(),
+    loadPackageDefinition: actual.loadPackageDefinition,
+    ServiceClientConstructor: actual.ServiceClientConstructor,
+    GrpcObject: actual.GrpcObject,
+  };
+});
+
+describe(`utils/grpc`, () => {
+  beforeEach(() => {
+    jest.clearAllMocks();
+  });
+
+  it(`should return a service client constructor`, () => {
+    const protoPath = path.resolve(__dirname, '../proto/proto/milvus.proto');
+    const proto = {
+      protoPath,
+      serviceName: `milvus.proto.milvus.MilvusService`,
+    };
+    const service = getGRPCService(proto);
+    expect(service).toBeDefined();
+  });
+
+  it(`should throw an error if the service object is invalid`, () => {
+    const protoPath = path.resolve(__dirname, '../proto/proto/milvus.proto');
+    const proto = {
+      protoPath,
+      serviceName: `milvus.proto.milvus.MilvusService2`,
+    };
+    expect(() => getGRPCService(proto)).toThrowError();
+  });
+
+  it('should add an authorization header to the metadata of a gRPC call', () => {
+    const username = 'testuser';
+    const password = 'testpassword';
+    const metadata = {
+      add: jest.fn(),
+    };
+    const listener = jest.fn();
+    const next = jest.fn();
+    const nextCall = jest.fn(() => ({
+      start: (metadata: any, listener: any, next: any) => {
+        next(metadata, listener);
+      },
+    }));
+    (InterceptingCall as any).mockImplementationOnce(
+      (call: any, options: any) => {
+        return {
+          call,
+          options,
+          start: options.start,
+        };
+      }
+    );
+
+    const interceptor = getAuthInterceptor(username, password);
+    const interceptedCall = interceptor({}, nextCall);
+
+    (interceptedCall.start as any)(metadata, listener, next);
+
+    expect(metadata.add).toHaveBeenCalledWith(
+      'authorization',
+      'dGVzdHVzZXI6dGVzdHBhc3N3b3Jk'
+    );
+  });
+});
